components: Fix LedMatrixMax truncating pin identifiers
Caused it to not work on platforms using more than 8bit for
pin ids, like the Mbed.

+++ microflo/components.cpp
@@ -1037,9 +1037,9 @@ private:
     }
 
 private:
-    int8_t pin_cs;
-    int8_t pin_din;
-    int8_t pin_clk;
+    int pin_cs;
+    int pin_din;
+    int pin_clk;
     bool initialized;
     uint8_t charIndex;
 };
